examples(audio_codecs): split into output_sdcard and output_speaker
Reorganize the audio examples so audio_codecs is a container with two
leaves that name their output path:
- output_speaker: the former audio_codecs (record then play back through
  a speaker/codec).
- output_sdcard: the former multimedia/audio_recorder/output_sdcard
  (record the mic to a WAV file on SD). Moved here and removed the now
  empty multimedia/audio_recorder.

Also: rename the recorder output to record.pcm/record.wav — ESP-IDF
FATFS defaults to 8.3 short names, so "tuya_recorder.wav" failed to
open; and add a SEEED_ESP32S3_XIAO_SENSE config for output_sdcard.
