--- examples/multimedia/audio_recorder/output_sdcard/app_default.config
@@ -1,1 +1,0 @@
-CONFIG_BOARD_CHOICE_T5AI=y
