--- examples/peripherals/audio_codecs/config/TUYA_T5AI_POCKET.config
@@ -1,2 +1,0 @@
-CONFIG_BOARD_CHOICE_T5AI=y
-CONFIG_BOARD_CHOICE_TUYA_T5AI_POCKET=y
