+++ examples/peripherals/audio_codecs/output_sdcard/config/SEEED_ESP32S3_XIAO_SENSE.config
@@ -1,0 +1,3 @@
+CONFIG_BOARD_CHOICE_ESP32=y
+CONFIG_BOARD_CHOICE_SEEED_ESP32S3_XIAO_SENSE=y
+CONFIG_ENABLE_SEEED_XIAO_SDCARD=y
