+++ examples/peripherals/audio_codecs/output_sdcard/config/TUYA_T5AI_POCKET.config
