+++ examples/peripherals/audio_codecs/output_speaker/CMakeLists.txt
