+++ examples/peripherals/audio_codecs/output_speaker/app_default.config
