+++ examples/peripherals/audio_codecs/output_speaker/config/DNESP32S3.config
