+++ examples/peripherals/audio_codecs/output_speaker/config/DNESP32S3_BOX.config
