+++ examples/peripherals/audio_codecs/output_speaker/config/ESP32S3_BREAD_COMPACT_WIFI.config
