+++ examples/peripherals/audio_codecs/output_speaker/config/RaspberryPi.config
