+++ examples/peripherals/audio_codecs/output_speaker/config/TUYA_T5AI_POCKET.config
