+++ examples/peripherals/audio_codecs/output_speaker/config/WAVESHARE_ESP32S3_Touch_AMOLED_1.8.config
