+++ examples/peripherals/audio_codecs/output_speaker/config/ZECTRIX_T5AI_NOTE_4.config
